HOTFIX - исправление ошибок в функционале favourites - Изменёны запросы в репозиториях. Добавлены ошибки, если у пользователя уже есть данный рецепт в избранном списке.

+++ src/main/java/com/miti/server/service/UserServiceImpl.java
@@ -164,20 +164,19 @@ public class UserServiceImpl implements UserService {
 
   @Override
   public void addFavouriteRecipe(Long userId, Long recipeId) {
-    try {
+    if (userRepository.getUserByFavouriteRecipe(userId, recipeId) == null)
       userRepository.addFavouriteRecipe(userId, recipeId);
-    } catch (Exception ex) {
-      ex.printStackTrace();
+    else {
+      throw new RuntimeException("");
     }
-
   }
 
   @Override
   public void deleteFavouriteRecipe(Long userId, Long recipeId) {
-    try {
+    if (userRepository.getUserByFavouriteRecipe(userId,recipeId) != null)
       userRepository.deleteFavouriteRecipe(userId, recipeId);
-    } catch (Exception ex) {
-      ex.printStackTrace();
+    else {
+      throw new RuntimeException("");
     }
   }
